+++ src/main/scala/edu/rit/cs/mmior/jsonoid/discovery/JsonoidParams.scala
@@ -1,20 +1,28 @@
 package edu.rit.cs.mmior.jsonoid.discovery
 
 final case class JsonoidParams(
+    val additionalProperties: Boolean,
     val er: EquivalenceRelation,
     val maxExamples: Int
 ) {
+  def withAdditionalProperties(
+      newAdditionalProperties: Boolean
+  ): JsonoidParams = {
+    JsonoidParams(newAdditionalProperties, er, maxExamples)
+  }
+
   def withER(newER: EquivalenceRelation): JsonoidParams = {
-    JsonoidParams(newER, maxExamples)
+    JsonoidParams(additionalProperties, newER, maxExamples)
   }
 
   def withMaxExamples(newMaxExamples: Int): JsonoidParams = {
-    JsonoidParams(er, newMaxExamples)
+    JsonoidParams(additionalProperties, er, newMaxExamples)
   }
 }
 
 object JsonoidParams {
   implicit val defaultJsonoidParams: JsonoidParams = JsonoidParams(
+    false,
     EquivalenceRelations.KindEquivalenceRelation,
     100
   )