+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add equivalence relation which checks types
 - Allow additional ER choices in CLI
 - Allow the maximum number of examples to be configured
+- Allow configuration of `additionalProperties`
 
 ### Changed
 - Use ranking to improve possible primary key suggestions